Create new class EvmManager
* Begin generalization of EthereumManager class.
* Move common methods to the new EvmManager class
* Added the following attributes to EvmManager: `etherscan_node`,
`etherscan_node_name`, `blockchain`, `contract_scan`, `contract_multicall`,
`contract_multicall_2`
* Renamed eth_rpc_timeout parameter to rpc_timeout
